search: add auth check to CreateCodeMonitor
We already have an auth check for edit operations (`isallowedtoedit`).
however, an auth check for "create" is a bit different because we don't
have to retrieve the owner of the monitor from the database to check
authorizations: we can simply take the owner from the input arguments.

I refactored `isAllowedToEdit` such that it shares most of the code with
`isAllowedToCreate`.